chore: Seperate LoadImage component
